+++ CHANGELOG.rst
@@ -8,6 +8,12 @@ Change Log
 
 5.4.0
 =====
+* Added support for check decorator boolean options: action_auto_run, action_manual_run, action_disable_run.
+  These will, respectively, force the associated action to run automatically after the check, or force it
+  not to run after the check (but allowed to run manually), or force it not to be allowed to run after the
+  check at all. These are mutually exclusive and if more than one are set (True) then they will all simply
+  be ignored. These work by setting the allow_action and the prevent_action properties of the check result,
+  and they will OVERRIDE these values if they were explicitly set within the check code itself.
 * Changes to the local-check-execute script.
 * Fix to checks.test_checks.add_random_test_nums.